pyproject.toml
[project]
name = "project-hack-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "google-genai>=1.13.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]==3.3.0",
    "sqlalchemy>=2.0.40",
]
